fix: time_limit type in ProblemEvent Object

+++ webapp/src/DataTransferObject/Shadowing/ProblemEvent.php
@@ -7,7 +7,7 @@ class ProblemEvent implements EventData
     public function __construct(
         public readonly string $id,
         public readonly string $name,
-        public readonly int $timeLimit,
+        public readonly float $timeLimit,
         public readonly ?string $label,
         public readonly ?string $rgb,
     ) {}
